requirements.txt
pytest==8.1.1
selenium==4.18.1
urllib3 == 2.2.1
allure-pytest==2.3.15
allure-python-commons==2.3.15
requests==2.31.0